Move AccountArguments to definitions module and document it
